pyproject.toml
[project]
name = "movies-newsletter-multi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "crewai>=0.67.1",
    "crewai-tools>=0.12.1",
    "langchain-google-genai>=1.0.4",
    "python-dotenv>=1.0.1",
    "pydantic>=2.9.2",
    "mlflow>=2.16.2",
    "jaraco-collections==5.1.0",
]

[project.optional-dependencies]
dev-dependency = [
    "ruff>=0.6.8",
]

[tool.uv]
dev-dependencies = [
    "jupyter>=1.1.1",
]

